refactor setup to share components and values and use new icons
