+++ apps/vbpay/app/setup/components/setup-form-steps.ts
@@ -1,0 +1,8 @@
+import { SteppedFormStep } from "@/types/stepped-form-step";
+import { Icons } from "@/components/icons";
+
+export const SetupFormSteps: SteppedFormStep[] = [
+  { id: 1, title: "License Information", icon: Icons.idCard },
+  { id: 2, title: "License Functionality", icon: Icons.squareFunction },
+  { id: 3, title: "Global Settings", icon: Icons.settings },
+];
